+++ firmware/boron-ins-fsm/src/consoleFunctions.cpp
@@ -16,8 +16,13 @@
 
 #include "Particle.h"
 #include "flashAddresses.h"
+<<<<<<< HEAD
 #include "im21door.h"
 #include "stateMachine.h"
+=======
+#include "stateMachine.h"
+#include "imDoorSensor.h"
+>>>>>>> 779f92df87eb9ef6fef4af4f56c80ed4f493b3bb
 
 void setupConsoleFunctions() {
   // particle console function declarations, belongs in setup() as per docs
@@ -47,8 +52,13 @@ bool isValidIM21Id(String input) {
   return true;
 }
 
+<<<<<<< HEAD
 int toggle_debugging_publishes(String command) {
   // default to invalid input
+=======
+int toggle_debugging_publishes(String command){
+  //default to invalid input
+>>>>>>> 779f92df87eb9ef6fef4af4f56c80ed4f493b3bb
   int returnFlag = -1;
 
   // string.toInt() returns 0 if it fails, so can't distinguish between user
@@ -58,7 +68,16 @@ int toggle_debugging_publishes(String command) {
   if (*(holder + 1) != 0) {
     // any string longer than 1 char is invalid input, so
     returnFlag = -1;
+<<<<<<< HEAD
   } else if (*holder == '0') {
+=======
+  }
+  // if e, echo whether debug publishes are on
+  else if (*holder == 'e') {
+    returnFlag = (int)stateMachineDebugFlag;
+  }
+  else if(*holder == '0'){
+>>>>>>> 779f92df87eb9ef6fef4af4f56c80ed4f493b3bb
     stateMachineDebugFlag = false;
     returnFlag = 0;
   } else if (*holder == '1') {
@@ -212,12 +231,22 @@ int im21_door_id_set(
   // get pointer to user-entered string
   const char* checkForEcho = command.c_str();
 
+<<<<<<< HEAD
   // if echo, publish current door ID
   if (*checkForEcho == 'e') {
     IM21DoorID holder;
     EEPROM.get(ADDR_IM21_DOORID, holder.byte1);
     EEPROM.get((ADDR_IM21_DOORID + 1), holder.byte2);
     EEPROM.get((ADDR_IM21_DOORID + 2), holder.byte3);
+=======
+  //if echo, publish current door ID
+  if(*checkForEcho == 'e'){
+
+    IMDoorID holder;   
+    EEPROM.get(ADDR_IM_DOORID,holder.byte1);  
+    EEPROM.get((ADDR_IM_DOORID+1),holder.byte2);  
+    EEPROM.get((ADDR_IM_DOORID+2),holder.byte3);  
+>>>>>>> 779f92df87eb9ef6fef4af4f56c80ed4f493b3bb
 
     char buffer[64];
     snprintf(buffer, sizeof(buffer),
@@ -231,6 +260,7 @@ int im21_door_id_set(
     const char* byteholder2;
     const char* byteholder3;
     int split1 = command.indexOf(',');
+<<<<<<< HEAD
     byteholder1 = command.substring(0, split1).c_str();
     globalDoorID.byte3 = (uint8_t)strtol(byteholder1, NULL, 16);
     int split2 = command.indexOf(',', split1 + 1);
@@ -246,6 +276,23 @@ int im21_door_id_set(
     EEPROM.put((ADDR_IM21_DOORID + 2), globalDoorID.byte3);
 
   }  // end if-else
+=======
+    byteholder1 = command.substring(0,split1).c_str();
+    globalDoorID.byte3 = (uint8_t)strtol(byteholder1,NULL,16);
+    int split2 = command.indexOf(',', split1+1);
+    byteholder2 = command.substring(split1+1,split2).c_str();
+    globalDoorID.byte2 = (uint8_t)strtol(byteholder2,NULL,16);
+    int split3 = command.indexOf(',', split2+1);
+    byteholder3 = command.substring(split2+1,split3).c_str();
+    globalDoorID.byte1 = (uint8_t)strtol(byteholder3,NULL,16);
+
+    //write new global door ID to flash
+    EEPROM.put(ADDR_IM_DOORID,globalDoorID.byte1);  
+    EEPROM.put((ADDR_IM_DOORID+1),globalDoorID.byte2);  
+    EEPROM.put((ADDR_IM_DOORID+2),globalDoorID.byte3);  
+  
+  } //end if-else
+>>>>>>> 779f92df87eb9ef6fef4af4f56c80ed4f493b3bb
 
   return 1;
 }