+++ firmware/boron-ins-fsm/test/ins3331Tests.cpp
@@ -1,28 +1,50 @@
 #define CATCH_CONFIG_MAIN
 #include "base.h"
+#include "../src/ins3331.cpp"
+#include "../src/ins3331.h"
+#include "../src/flashAddresses.h"
 
-unsigned char calculateChecksum(unsigned char myArray[], int arrayLength){
+SCENARIO("The correct checksum is returned by calculateChecksum()")
+{
+    GIVEN("The buffer is populated with zeros")
+    {
+        unsigned char buffer[15] = {0};
 
-  unsigned char checksum = 0x00;
-  
-  for(int i = 2; i <= 12; i++){
-    checksum = checksum + myArray[i];
-  }
-  
-  return checksum;
-
-}
-
-SCENARIO("The correct checksum is returned by calculateChecksum()") {
-
-    GIVEN("The buffer is populated with zeros") {
-        unsigned char buffer[15] = { 0 };
-
-        WHEN("The function is called") {
+        WHEN("The function is called")
+        {
             int checksum = calculateChecksum(buffer, 15);
 
-            THEN("Checksum should be equal to 0") {
-                REQUIRE(checksum == 0);
+            THEN("Checksum should be equal to 0") { 
+                REQUIRE(checksum == 0); 
+            }
+        }
+    }
+    GIVEN("The buffer is populated with ones")
+    {
+        unsigned char buffer[15];
+        for (int i = 0; i < 15; ++i) {
+            buffer[i] = 1;
+        }
+
+        WHEN("The function is called")
+        {
+            int checksum = calculateChecksum(buffer, 15);
+
+            THEN("Checksum should be equal to 11") { 
+                REQUIRE(checksum == 11); 
+            }
+        }
+    }
+    GIVEN("The buffer is populated with alternating zeros and ones, starting with zero")
+    {
+        unsigned char buffer[15] = {0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0};
+
+        WHEN("The function is called")
+        {
+            int checksum = calculateChecksum(buffer, 15);
+
+            THEN("Checksum should be equal to 5") { 
+                REQUIRE(checksum == 5); 
             }
         }
     }