+++ firmware/boron-ins-fsm/test/mocks/Particle.h
@@ -1,10 +1,26 @@
+/*
+ * Mock implementations for Particle.h library functions.
+ 
+ * The mock implementations are necessary so that unit tests
+ * can be compiled using GCC and run automatically in Travis.
+ */
+
 #pragma once
+
 #include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <time.h>
 #include <functional>
+
 #include "../../inc/spark_wiring_string.h"
+#include "helper.h"
+#include "mock_serial.h"
+#include "mock_thread.h"
+
+uint32_t millis();
+void delay(unsigned long ms);
 
 String fullPublishString;
 enum PublishFlag
@@ -15,10 +31,8 @@ enum PublishFlag
     WITH_ACK
 };
 
-struct os_queue_t
-{
-
-}; 
+// Fake structs
+struct os_queue_t {};
 
 class MockParticle
 {