+++ server/stateMachine/SessionStateEnum.js
@@ -1,8 +1,0 @@
-const STATE = {
-  IDLE: 'Idle',
-  INITIAL_TIMER: 'Initial Timer',
-  DURATION_TIMER: 'Duration Timer',
-  STILLNESS_TIMER: 'Stillness Timer',
-}
-
-module.exports = Object.freeze(STATE)